[step 3] refactors : 리뷰 반영
* 총 개수 부분 버그 수정
* utilFunc.js 파일 2개로 분리
* 반복문을 map 메서드로 변경
