+++ src/utils/renderer.js
@@ -1,17 +1,19 @@
 import { $ } from "./selector.js";
 
 export const updateLottoAmounts = (selector, amounts) => {
-  $(selector).innerHTML = `
-      <label class="flex-auto my-0 lotto-amounts" data-cy="lotto-amounts"
-      >총 ${amounts}개를 구매하였습니다.</label
-      >
-      <div class="flex-auto d-flex justify-end pr-1">
-        <label class="switch">
-          <input type="checkbox" class="lotto-numbers-toggle-btn" data-cy="lotto-numbers-toggle-btn" />
-          <span class="text-base font-normal">번호보기</span>
-        </label>
-      </div>
-    `;
+  const template = `
+  <label class="flex-auto my-0 lotto-amounts" data-cy="lotto-amounts"
+  >총 ${amounts}개를 구매하였습니다.</label
+  >
+  <div class="flex-auto d-flex justify-end pr-1">
+    <label class="switch">
+      <input type="checkbox" class="lotto-numbers-toggle-btn" data-cy="lotto-numbers-toggle-btn" />
+      <span class="text-base font-normal">번호보기</span>
+    </label>
+  </div>
+`;
+
+  $(selector).innerHTML = template;
 };
 
 export const updateLottoTickets = (amounts, lottoTicketsList) => {
@@ -24,6 +26,7 @@ export const updateLottoTickets = (amounts, lottoTicketsList) => {
         </span>
       </div>`
   );
+
   $(".lotto__tickets").innerHTML = newLottoTickets.join("");
 };
 
@@ -38,11 +41,9 @@ export const updateManualLottoTickets = (amounts, lottoTicketsList) => {
       </div>
     `;
 
-  let newLottoTickets = "";
-  for (let i = 0; i < amounts; i++) {
-    newLottoTickets += template(i);
-  }
-  $(".manual-lotto__tickets").innerHTML = newLottoTickets;
+  $(".manual-lotto__tickets").innerHTML = Array.from(Array(amounts)).map(
+    (index) => template(index)
+  );
 };
 
 export const resetLotto = () => {
@@ -98,11 +99,12 @@ export const addManualNumberInput = (selector, amounts) => {
       </div>
     </div>`;
 
-  // Array.map이 안먹힌다..
   let newManualNumberInputs = [];
+
   for (let i = 0; i < amounts; i++) {
     newManualNumberInputs.push(template(i));
   }
+
   $(selector).innerHTML = `
     <label class="flex-auto d-inline-block mb-3">로또 번호를 입력해주세요.</label>
     <div class="d-flex flex-col">
@@ -119,61 +121,62 @@ export const addManualNumberInput = (selector, amounts) => {
 };
 
 export const addWinningNumberInput = (selector) => {
-  $(selector).innerHTML = `
-    <label class="flex-auto d-inline-block mb-3"
-    >지난 주 당첨번호 6개와 보너스 넘버 1개를 입력해주세요.</label
-      >
-      <div class="d-flex">
+  const template = `
+  <label class="flex-auto d-inline-block mb-3"
+  >지난 주 당첨번호 6개와 보너스 넘버 1개를 입력해주세요.</label
+    >
+    <div class="d-flex">
+      <div>
+        <h4 class="mt-0 mb-3 text-center">당첨 번호</h4>
         <div>
-          <h4 class="mt-0 mb-3 text-center">당첨 번호</h4>
-          <div>
-            <input
-              type="number"
-              class="winning-number mx-1 text-center"
-              data-cy="winning-number-1"
-            />
-            <input
-              type="number"
-              class="winning-number mx-1 text-center"
-              data-cy="winning-number-2"
-            />
-            <input
-              type="number"
-              class="winning-number mx-1 text-center"
-              data-cy="winning-number-3"
-            />
-            <input
-              type="number"
-              class="winning-number mx-1 text-center"
-              data-cy="winning-number-4"
-            />
-            <input
-              type="number"
-              class="winning-number mx-1 text-center"
-              data-cy="winning-number-5"
-            />
-            <input
-              type="number"
-              class="winning-number mx-1 text-center"
-              data-cy="winning-number-6"
-            />
-          </div>
-        </div>
-        <div class="bonus-number-container flex-grow">
-          <h4 class="mt-0 mb-3 text-center">보너스 번호</h4>
-          <div class="d-flex justify-center">
-            <input type="number" class="bonus-number text-center" data-cy="bonus-number"/>
-          </div>
+          <input
+            type="number"
+            class="winning-number mx-1 text-center"
+            data-cy="winning-number-1"
+          />
+          <input
+            type="number"
+            class="winning-number mx-1 text-center"
+            data-cy="winning-number-2"
+          />
+          <input
+            type="number"
+            class="winning-number mx-1 text-center"
+            data-cy="winning-number-3"
+          />
+          <input
+            type="number"
+            class="winning-number mx-1 text-center"
+            data-cy="winning-number-4"
+          />
+          <input
+            type="number"
+            class="winning-number mx-1 text-center"
+            data-cy="winning-number-5"
+          />
+          <input
+            type="number"
+            class="winning-number mx-1 text-center"
+            data-cy="winning-number-6"
+          />
         </div>
       </div>
-      <button
-        type="button"
-        class="open-result-modal-button mt-5 btn btn-cyan w-100"
-        data-cy="result-modal-btn"
-      >
-        결과 확인하기
-      </button>
-    `;
+      <div class="bonus-number-container flex-grow">
+        <h4 class="mt-0 mb-3 text-center">보너스 번호</h4>
+        <div class="d-flex justify-center">
+          <input type="number" class="bonus-number text-center" data-cy="bonus-number"/>
+        </div>
+      </div>
+    </div>
+    <button
+      type="button"
+      class="open-result-modal-button mt-5 btn btn-cyan w-100"
+      data-cy="result-modal-btn"
+    >
+      결과 확인하기
+    </button>
+  `;
+  $(selector).innerHTML = template;
 };
 
 export const updateResult = (lottoResultCount, lottoResult) => {
